Add Terraform file to run acceptance tests
The acceptance tests are difficult to run as they require an OCI
instance and user configured in a very specific way.
This simplifies the process of configuring OCI and running the
acceptance tests by packaging the setup in a Terraform file.

Also updates the README with information on the acceptance tests.